+++ src/renderer/src/components/message/MessageBlock.tsx
@@ -29,6 +29,8 @@ import {
   TASKS_TOOL_GET_TASK_MESSAGE,
   TASKS_TOOL_CREATE_TASK,
   TASKS_TOOL_DELETE_TASK,
+  TASKS_TOOL_SEARCH_TASK,
+  TASKS_TOOL_SEARCH_PARENT_TASK,
 } from '@common/tools';
 
 import { CommandOutputMessageBlock } from './CommandOutputMessageBlock';
@@ -52,6 +54,8 @@ import { GetTaskToolMessage } from './GetTaskToolMessage';
 import { GetTaskMessageToolMessage } from './GetTaskMessageToolMessage';
 import { CreateTaskToolMessage } from './CreateTaskToolMessage';
 import { DeleteTaskToolMessage } from './DeleteTaskToolMessage';
+import { SearchTaskToolMessage } from './SearchTaskToolMessage';
+import { SearchParentTaskToolMessage } from './SearchParentTaskToolMessage';
 import { StoreMemoryToolMessage } from './StoreMemoryToolMessage';
 import { RetrieveMemoryToolMessage } from './RetrieveMemoryToolMessage';
 import { DeleteMemoryToolMessage } from './DeleteMemoryToolMessage';
@@ -175,6 +179,10 @@ const MessageBlockComponent = ({ baseDir, taskId, message, allFiles, renderMarkd
             return <CreateTaskToolMessage message={toolMessage} onRemove={remove} compact={compact} />;
           case TASKS_TOOL_DELETE_TASK:
             return <DeleteTaskToolMessage message={toolMessage} onRemove={remove} compact={compact} />;
+          case TASKS_TOOL_SEARCH_TASK:
+            return <SearchTaskToolMessage message={toolMessage} onRemove={remove} compact={compact} />;
+          case TASKS_TOOL_SEARCH_PARENT_TASK:
+            return <SearchParentTaskToolMessage message={toolMessage} onRemove={remove} compact={compact} />;
           default:
             break;
         }